+++ tests/Properties/MkShow.hs
@@ -24,7 +24,6 @@ import           Derived
 
 import           Instances.Derived ()
 
-import           Test.Tasty.QuickCheck (NonZero)
 import           Test.Tasty (TestTree, testGroup)
 import           Test.Tasty.QuickCheck (Arbitrary, testProperty)
 
@@ -56,9 +55,9 @@ prop_mkShow s1 s2 s3 sp1 sp2 sp3 sl1 sl2 sl3 p a =
     && fromString (showList    [a] "") == sl3   [a]
 
 -- | Verifies 'mkShow' (and related functions) produce the same output as their
--- 'String' counterparts. This uses a data type that is a 'Show' instance.
-prop_mkShowIsInstance :: Int -> AllAtOnce Int Int Int Int -> Bool
-prop_mkShowIsInstance = prop_mkShow
+-- 'String' counterparts. This uses a plain type constructor.
+prop_mkShowTyCon :: Int -> AllAtOnce Int Int Int Int -> Bool
+prop_mkShowTyCon = prop_mkShow
     $(mkShow         ''AllAtOnce)
     $(mkShowLazy     ''AllAtOnce)
     $(mkShowb        ''AllAtOnce)
@@ -69,31 +68,49 @@ prop_mkShowIsInstance = prop_mkShow
     $(mkShowListLazy ''AllAtOnce)
     $(mkShowbList    ''AllAtOnce)
 
+#if MIN_VERSION_template_haskell(2,7,0)
 -- | Verifies 'mkShow' (and related functions) produce the same output as their
--- 'String' counterparts. This uses a data type that is not a 'Show' instance.
-prop_mkShowIsNotInstance :: Int -> NonZero Int -> Bool
-prop_mkShowIsNotInstance = prop_mkShow
-    $(mkShow         ''NonZero)
-    $(mkShowLazy     ''NonZero)
-    $(mkShowb        ''NonZero)
-    $(mkShowPrec     ''NonZero)
-    $(mkShowPrecLazy ''NonZero)
-    $(mkShowbPrec    ''NonZero)
-    $(mkShowList     ''NonZero)
-    $(mkShowListLazy ''NonZero)
-    $(mkShowbList    ''NonZero)
+-- 'String' counterparts. This uses a data family name.
+prop_mkShowDataFam :: Int -> OneDataInstance Int Int Int Int -> Bool
+prop_mkShowDataFam = prop_mkShow
+    $(mkShow         ''OneDataInstance)
+    $(mkShowLazy     ''OneDataInstance)
+    $(mkShowb        ''OneDataInstance)
+    $(mkShowPrec     ''OneDataInstance)
+    $(mkShowPrecLazy ''OneDataInstance)
+    $(mkShowbPrec    ''OneDataInstance)
+    $(mkShowList     ''OneDataInstance)
+    $(mkShowListLazy ''OneDataInstance)
+    $(mkShowbList    ''OneDataInstance)
 
--- prop_mkPrintIsInstance
--- prop_mkPrintIsNotInstance
--- prop_traceIsInstance
--- prop_traceIsNotInstance
+-- | Verifies 'mkShow' (and related functions) produce the same output as their
+-- 'String' counterparts. This uses a data family instance constructor.
+prop_mkShowDataFamInstCon :: Int -> NotAllShow Int Int Int Int -> Bool
+prop_mkShowDataFamInstCon = prop_mkShow
+    $(mkShow         'NASShow1)
+    $(mkShowLazy     'NASShow1)
+    $(mkShowb        'NASShow1)
+    $(mkShowPrec     'NASShow1)
+    $(mkShowPrecLazy 'NASShow1)
+    $(mkShowbPrec    'NASShow1)
+    $(mkShowList     'NASShow1)
+    $(mkShowListLazy 'NASShow1)
+    $(mkShowbList    'NASShow1)
+#endif
+
+-- prop_mkPrint
+-- prop_trace
 
 mkShowTests :: [TestTree]
 mkShowTests =
     [ testGroup "mkShow and related functions"
-        [ testProperty "$(mkShow ''AllAtOnce) (a Show instance)"    prop_mkShowIsInstance
-        , testProperty "$(mkShow ''NonZero) (not a Show instance)"  prop_mkShowIsNotInstance
---         , testProperty "$(mkPrint ''AllAtOnce) (a Show instance)"   prop_mkPrintIsInstance
---         , testProperty "$(mkPrint ''NonZero) (not a Show instance)" prop_mkPrintIsNotInstance
+        [ testProperty "$(mkShow ''AllAtOnce) (a plain type constructor)"             prop_mkShowTyCon
+--         , testProperty "$(mkPrint ''AllAtOnce) (a plain type constructor)"            prop_mkPrintTyCon
+#if MIN_VERSION_template_haskell(2,7,0)
+        , testProperty "$(mkShow ''NotAllShow) (a data family instance constructor)"  prop_mkShowDataFamInstCon
+--         , testProperty "$(mkPrint ''NotAllShow) (a data family instance constructor)" prop_mkShowDataFamInstCon
+        , testProperty "$(mkShow ''OneDataInstance) (a data family name)"             prop_mkShowDataFam
+--         , testProperty "$(mkPrint ''OneDataInstance) (a data family name)"            prop_mkPrintDataFam
+#endif
         ]
     ]